Replaced PrismaClientRust with SQLX

+++ src/ai.rs
@@ -5,12 +5,12 @@ use lazy_static::lazy_static;
 use crate::{services::{game_service::models::Stats, battle_service::data_layer::{ATTACK_IDX, DEFEND_IDX, IDLE_IDX}}, dice};
 
 pub trait Ai : Send + Sync {
-    fn next_act(&self, pl_stats: &Stats, monst_stats: &Stats) -> i32;
+    fn next_act(&self, pl_stats: &Stats, monst_stats: &Stats) -> i64;
 }
 
 struct WilfredAi;
 impl Ai for WilfredAi {
-    fn next_act(&self, _pl_stats: &Stats, monst_stats: &Stats) -> i32 {
+    fn next_act(&self, _pl_stats: &Stats, monst_stats: &Stats) -> i64 {
         if monst_stats.power < 2 {
             return if dice::single(2) == 1 { DEFEND_IDX } else { IDLE_IDX };
         }
@@ -20,7 +20,7 @@ impl Ai for WilfredAi {
 
 struct NutcrackerAi;
 impl Ai for NutcrackerAi {
-    fn next_act(&self, _pl_stats: &Stats, _monst_stats: &Stats) -> i32 {
+    fn next_act(&self, _pl_stats: &Stats, _monst_stats: &Stats) -> i64 {
         match dice::single(3) {
             1 => IDLE_IDX,
             2 => ATTACK_IDX,
